+++ Projects/IRMA Client/WholeFoods.IRMA Solution/IRMA/ItemHosting/UserInterface/Form_ItemScaleDetails365.vb
@@ -1566,10 +1566,10 @@ Imports log4net
         scaleNutriFact.Chromium = CInt(ChromiumNumericEditor.Value.ToString())
         scaleNutriFact.Copper = CInt(CopperNumericEditor.Value.ToString())
         scaleNutriFact.Description = NutrifactDescriptionTextbox.Text
-        scaleNutriFact.MonounsaturatedFat = CInt(FatMonoNumericEditor.Value.ToString())
+        scaleNutriFact.MonounsaturatedFat = CDec(FatMonoNumericEditor.Value.ToString())
         scaleNutriFact.Om3Fatty = FatOmega3NumericEditor.Value.ToString()
         scaleNutriFact.Om6Fatty = FatOmega6NumericEditor.Value.ToString()
-        scaleNutriFact.PolyunsaturatedFat = CInt(FatPolyNumericEditor.Value.ToString())
+        scaleNutriFact.PolyunsaturatedFat = CDec(FatPolyNumericEditor.Value.ToString())
         scaleNutriFact.SaturatedFatWeight = FatSatNumericEditor.Value.ToString()
         scaleNutriFact.SaturatedFatPercent = CInt(FatSatPercentNumericEditor.Value.ToString())
         scaleNutriFact.TotalFatWeight = FatTotalNumericEditor.Value.ToString()
